JOLTEON-10 Changing Bike entity and respective tests

+++ backend/src/main/java/tqs/project/jolteon/entities/Bike.java
@@ -13,8 +13,8 @@ public class Bike {
     private Long id;
 
     private Float battery;
-    private Long autonomy;
+    private Integer autonomy;
     private Boolean isAvailable;
-    private Float latitude;
-    private Float longitude;
+    private Double latitude;
+    private Double longitude;
 }
